pedantic ansi support

+++ init_iter.c
@@ -12,11 +12,11 @@
 
 #include "ftlst.h"
 #include <stdlib.h>
-#include <stdint.h>
+#include <limits.h>
 
 void	init_iter(t_lstiter *it, t_lst *l, t_lst_direction d)
 {
-	it->pos = d == increasing ? SIZE_MAX : lst_get_size(l);
+	it->pos = d == increasing ? INT_MAX : lst_get_size(l);
 	it->current = d == increasing ? lst_node_back(l) : lst_node_front(l);
 	it->end = d == increasing ? lst_node_front(l) : lst_node_back(l);
 	it->dir = d;
